removed position rounding in library

+++ gwt-connectors/src/main/java/pl/tecna/gwt/connectors/client/CornerPoint.java
@@ -2,7 +2,7 @@ package pl.tecna.gwt.connectors.client;
 
 public class CornerPoint extends Point {
 
-  public CornerPoint(Integer left, Integer top) {
+  public CornerPoint(double left, double top) {
     super(left, top);
   }
